Testes Unitários e Correções(Cliente, Cartao, Fornecedor)

+++ placadevideo-tp1-quarkus/src/main/java/br/unitins/tp1/placadevideo/dto/PlacaDeVideoRequestDTO.java
@@ -11,7 +11,7 @@ public record PlacaDeVideoRequestDTO(
         @NotBlank(message = "O campo deve ser informado")
         String categoria,
         @NotBlank(message = "O campo deve ser informado")
-        double preco,
+        float preco,
         @NotBlank(message = "O campo deve ser informado")
         int vram,
         @NotBlank(message = "O campo deve ser informado")
