+++ placadevideo-tp1-quarkus/src/main/java/br/unitins/tp1/placadevideo/dto/PlacaDeVideoResponseDTO.java
@@ -6,7 +6,7 @@ public record PlacaDeVideoResponseDTO(
         Long id,
         String modelo,
         String categoria,
-        double preco,
+        float preco,
         int vram,
         String resolucao,
         int energia,
